+++ src/Features.CSharp/NetObjectsModel.cs
