+++ src/Features.CSharp/SignalsModel.cs
