+++ src/Features.Common/Features.Common.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+  
+  <ItemGroup>
+    <Content Include="images\**" CopyToPublishDirectory="Always" CopyToOutputDirectory="Always" />
+    <Content Include="pages\**" CopyToPublishDirectory="Always" CopyToOutputDirectory="Always" />
+    <Content Include="Main.qml" CopyToPublishDirectory="Always" CopyToOutputDirectory="Always" />
+  </ItemGroup>
+
+</Project>
