+++ src/Features.Common/images/qt-logo.png
