+++ src/Features.Common/pages/Dynamics.qml
