+++ src/Features.FSharp/NetObjectsModel.fs
@@ -1,0 +1,8 @@
+namespace Features
+
+type NetObject() =
+    member val Property = "" with get, set
+    member this.UpdateProperty(value: string) = this.Property <- value
+
+type NetObjectsModel() =
+    member this.GetNetObject() = NetObject()
